requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
attrs=23.2.0=pypi_0
ca-certificates=2023.12.12=h06a4308_0
certifi=2024.2.2=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
click-plugins=1.1.1=pypi_0
cligj=0.7.2=pypi_0
cloudpickle=3.0.0=pypi_0
coloredlogs=15.0.1=pypi_0
contourpy=1.2.0=pypi_0
cycler=0.12.1=pypi_0
exceptiongroup=1.2.0=pypi_0
farama-notifications=0.0.4=pypi_0
fiona=1.9.5=pypi_0
flatbuffers=23.5.26=pypi_0
fonttools=4.47.2=pypi_0
geopandas=0.14.3=pypi_0
gymnasium=0.28.1=pypi_0
humanfriendly=10.0=pypi_0
idna=3.6=pypi_0
importlib-resources=6.1.1=pypi_0
iniconfig=2.0.0=pypi_0
jax-jumpy=1.0.0=pypi_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
lxml=5.1.0=pypi_0
matplotlib=3.8.2=pypi_0
metadrive-simulator=0.4.2.3=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.4=h6a678d5_0
numpy=1.24.2=pypi_0
onnxruntime-gpu=1.17.0=pypi_0
openssl=3.0.13=h7f8727e_0
packaging=23.2=pypi_0
panda3d=1.10.13=pypi_0
panda3d-gltf=0.13=pypi_0
panda3d-simplepbr=0.11.2=pypi_0
pandas=2.2.0=pypi_0
pillow=10.2.0=pypi_0
pip=23.3.1=py39h06a4308_0
pluggy=1.4.0=pypi_0
progressbar=2.5=pypi_0
protobuf=4.25.2=pypi_0
psutil=5.9.8=pypi_0
pygame=2.5.2=pypi_0
pygments=2.17.2=pypi_0
pyparsing=3.1.1=pypi_0
pyproj=3.6.1=pypi_0
pytest=8.0.0=pypi_0
python=3.9.18=h955ad1f_0
python-dateutil=2.8.2=pypi_0
pytz=2024.1=pypi_0
readline=8.2=h5eee18b_0
requests=2.31.0=pypi_0
scipy=1.12.0=pypi_0
seaborn=0.13.2=pypi_0
setuptools=68.2.2=py39h06a4308_0
shapely=2.0.2=pypi_0
six=1.16.0=pypi_0
sqlite=3.41.2=h5eee18b_0
sympy=1.12=pypi_0
tk=8.6.12=h1ccaba5_0
tqdm=4.66.1=pypi_0
tzdata=2023.4=pypi_0
urllib3=2.2.0=pypi_0
wheel=0.41.2=py39h06a4308_0
xz=5.4.5=h5eee18b_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
